pyproject.toml
[project]
name = "risk-agent-sber"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.11.18",
    "aiosignal==1.3.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "appdirs==1.4.4",
    "attrs==25.3.0",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.1",
    "dataclasses-json==0.6.7",
    "eval-type-backport==0.2.2",
    "filelock==3.18.0",
    "frozenlist==1.6.0",
    "fsspec==2025.3.2",
    "greenlet==3.2.1",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httpx-sse==0.4.0",
    "huggingface-hub==0.30.2",
    "idna==3.10",
    "importlib-metadata==8.7.0",
    "jinja2==3.1.6",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain==0.3.24",
    "langchain-community==0.3.23",
    "langchain-core==0.3.56",
    "langchain-mistralai==0.2.10",
    "langchain-nomic>=0.1.4",
    "langchain-text-splitters==0.3.8",
    "langgraph==0.4.1",
    "langgraph-checkpoint==2.0.25",
    "langgraph-prebuilt==0.1.8",
    "langgraph-sdk==0.1.66",
    "langsmith==0.3.39",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==3.26.1",
    "mdurl==0.1.2",
    "mistralai==1.7.0",
    "multidict==6.4.3",
    "mypy-extensions==1.1.0",
    "nest-asyncio==1.6.0",
    "numpy==2.2.5",
    "orjson==3.10.18",
    "ormsgpack==1.9.1",
    "packaging==24.2",
    "propcache==0.3.1",
    "pydantic==2.11.4",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.9.1",
    "pyee==11.1.1",
    "pygments==2.19.1",
    "pypdf2==3.0.1",
    "pyppeteer==2.0.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==14.0.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.40",
    "tavily-python==0.7.0",
    "tenacity==9.1.2",
    "tiktoken==0.9.0",
    "tokenizers==0.21.1",
    "tqdm==4.67.1",
    "typing==3.10.0.0",
    "typing-extensions==4.13.2",
    "typing-inspect==0.9.0",
    "typing-inspection==0.4.0",
    "urllib3==1.26.20",
    "websockets==10.4",
    "xxhash==3.5.0",
    "yarl==1.20.0",
    "zipp==3.21.0",
    "zstandard==0.23.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"
